fix: Improve authenticated guard
