--- src/features/auth/hooks.ts
@@ -1,8 +1,0 @@
-import { trpc } from '@/lib/trpc/client';
-
-export const useCheckAuthenticated = () => {
-  return trpc.auth.checkAuthenticated.useQuery(undefined, {
-    staleTime: 30000,
-    cacheTime: Infinity,
-  });
-};
