+++ src/server/routers/auth.tsx
@@ -12,7 +12,7 @@ import {
   VALIDATION_RETRY_DELAY_IN_SECONDS,
   VALIDATION_TOKEN_EXPIRATION_IN_MINUTES,
 } from '@/features/auth/utils';
-import { zUser } from '@/features/users/schemas';
+import { zUser, zUserAuthorization } from '@/features/users/schemas';
 import i18n from '@/lib/i18n/server';
 import {
   AUTH_COOKIE_NAME,
@@ -34,12 +34,18 @@ export const authRouter = createTRPCRouter({
       },
     })
     .input(z.void())
-    .output(z.object({ isAuthenticated: z.boolean() }))
+    .output(
+      z.object({
+        isAuthenticated: z.boolean(),
+        authorizations: z.array(zUserAuthorization()).optional(),
+      })
+    )
     .query(async ({ ctx }) => {
       ctx.logger.info(`User ${ctx.user ? 'is' : 'is not'} logged`);
 
       return {
         isAuthenticated: !!ctx.user,
+        authorizations: ctx.user?.authorizations,
       };
     }),
